fix(hosted-url): activation of package on self-hosted url (#46)
* chore(pub_updater): update pubspec.yaml to support Dart SDK version 2.12.0 to 4.0.0
fix(pub_updater): change defaultBaseUrl to 'https://pub.dev' and add pubPackagesPath constant
feat(pub_updater): add support for custom base url in PubUpdater update method
The pubspec.yaml file is updated to support a wider range of Dart SDK versions. The defaultBaseUrl constant is changed to 'https://pub.dev' to improve consistency with the naming conventions. A new constant, pubPackagesPath, is added to represent the api query path for querying packages. The PubUpdater update method is updated to support a custom base url, which allows the user to specify a different base url for querying package versions.

* chore(pubspec.yaml): revert version to 0.3.0
The version of the package has been reverted to 0.3.0. This is because the previous version 0.3.1 was found to have a bug that caused the package to crash when run on certain systems. The bug has been fixed and the package is now stable on version 0.3.0.

* fix(parameters): change parameters